Fix double ended iterator issue and refactor waveform module
