+++ src/fft.rs
@@ -3,21 +3,17 @@ pub mod czt;
 pub mod complex;
 use num::Integer;
 use num_complex::Complex;
-use num_traits::{Float, FloatConst, NumAssign, AsPrimitive, NumAssignOps};
-use std::ops::{IndexMut, Deref};
+use num_traits::{ Float, FloatConst, NumAssign, AsPrimitive, NumAssignOps };
+use std::ops::IndexMut;
 use crate::traits::Iterable;
 
 pub fn dtft<F, I, C>(x: I) -> impl Fn(I) -> C
 where 
-    // Bound F to float types
     F: Float + FloatConst + NumAssign + 'static,
-    // Bound I to to an iterable collection of F
     I: FromIterator<F> + Clone,
     for<'a> &'a I: IntoIterator<Item = &'a F>,
     for<'a> <&'a I as IntoIterator>::IntoIter: ExactSizeIterator,
-    // Ensure a usize can be converted to F, ideally this can be removed
     usize: AsPrimitive<F>,
-    // Bound C to a collection of Complex<F>
     C: FromIterator<Complex<F>> + IndexMut<usize, Output = Complex<F>>,
 {
     move |samples: I| -> C {
@@ -33,34 +29,20 @@ where
 
 pub fn idft<F, I, C>(x: &I) -> C
 where
-    // Bound F to float types
     F: Float + FloatConst + NumAssign + 'static,
     for<'c> I: Iterable<OwnedItem = Complex<F>, Item<'c> = &'c Complex<F>>,
-    //for<'c> I: 'c, for<'c> C: 'c,
-    //for<'c> <I as Iterable>::Item<'c>: Deref<Target = F>,
     for<'c> C: Iterable<OwnedItem = F, Item<'c> = &'c F>,
-    // Bound I to to an iterable collection of F
-    //I: FromIterator<Complex<F>> + Clone,
-    //for<'a> &'a I: IntoIterator<Item = &'a Complex<F>>,
-    //for<'a> <&'a I as IntoIterator>::IntoIter: ExactSizeIterator,
-    // Ensure a usize can be converted to F, ideally this can be removed
     usize: AsPrimitive<F>,
-    // Bound C to a collection of Complex<F>
-    //C: FromIterator<F> + IndexMut<usize, Output = F>,
 {
     complex::idft_internal(x).iter().map(|x| x.re).collect()
 }
 
 pub fn dft<F, I, C>(x: &I) -> C
 where
-    // Bound F to float types
     F: Float + FloatConst + NumAssign + 'static,
     for<'c> I: Iterable<OwnedItem = F, Item<'c> = &'c F>,
-
-    //for<'c> <I as Iterable>::Item<'c>: Deref<Target = F>,
-    // Ensure a usize can be converted to F, ideally this can be removed
-    usize: AsPrimitive<F>,
     for<'c> C: Iterable<OwnedItem = Complex<F>, Item<'c> = &'c Complex<F>>,
+    usize: AsPrimitive<F>,
 {
     let n = x.len();
     let zero = F::zero();
@@ -75,13 +57,9 @@ where
 
 pub fn fftfreq<F, I>(n: usize, d: F) -> I
 where
-    // Bound F to float types
     F: Float + FloatConst + NumAssign + 'static,
     usize: AsPrimitive<F>,
     for<'c> I: Iterable<OwnedItem = F, Item<'c> = &'c F>,
-    //for<'c> I: 'c,
-    // Bound I to to an iterable collection of F
-    //I: FromIterator<F> + Clone,
 {
     let time = d * n.as_();
     (0..n).map(|i| i.as_() / time).collect()
@@ -89,14 +67,10 @@ where
 
 pub fn fftfreq_balanced<F, I>(n: usize, d: F) -> I 
 where
-    // Bound F to float types
     F: Float + FloatConst + NumAssign + 'static,
     usize: AsPrimitive<F>,
     i32: AsPrimitive<F>,
-    // Bound I to to an iterable collection of F
     for<'c> I: Iterable<OwnedItem = F, Item<'c> = &'c F>,
-    //for<'c> I: 'c,
-    //I: FromIterator<F> + Clone,
 {
     let time = d * n.as_();
     let (pos_end, neg_start) = if n.is_even() {
