+++ src/fft/ct/complex.rs
@@ -6,12 +6,9 @@ use crate::traits::Iterable;
 
 pub fn fft<F, I>(x: &I) -> I
 where
-    // Bound F to float types
     F: Float + FloatConst + NumAssign + 'static,
-    // Bound I to to an iterable collection of F
     for<'c> I: Iterable<OwnedItem = Complex<F>, Item<'c> = &'c Complex<F>>,
     I: IndexMut<usize, Output = Complex<F>>,
-    // Ensure a usize can be converted to F, ideally this can be removed
     usize: AsPrimitive<F>,
 {
     let n = x.len();
@@ -48,15 +45,9 @@ where
 
 pub(crate) fn ifft_internal<F, I>(x: &I) -> I
 where
-    // Bound F to float types
     F: Float + FloatConst + NumAssign + 'static,
     for<'c> I: Iterable<OwnedItem = Complex<F>, Item<'c> = &'c Complex<F>>,
     I: IndexMut<usize, Output = Complex<F>>,
-    // Bound I to an iterable collection of F
-    //I: FromIterator<Complex<F>> + Clone + IndexMut<usize, Output = Complex<F>>,
-    //for<'a> &'a I: IntoIterator<Item = &'a Complex<F>>,
-    //for<'a> <&'a I as IntoIterator>::IntoIter: ExactSizeIterator,
-    // Ensure a usize can be converted to F
     usize: AsPrimitive<F>,
 {
     let n = x.len();
